+++ documentation/ridge_example.py
@@ -62,8 +62,8 @@ W_in = input_connections(N, m, p_in, variance=Delta_in, zero_mean=True)
 
 # initialize network
 net = Network.from_yaml("neuron_model_templates.spiking_neurons.qif.qif_sfa_pop", weights=W*J,
-                        source_var="s", target_var="s_in", input_var_ext="I_ext", output_var="s", spike_var="v",
-                        input_var_net="spike", op="qif_sfa_op", node_vars={'all/qif_sfa_op/eta': etas}, dt=dt,
+                        source_var="s", target_var="s_in", input_var="I_ext", output_var="s", spike_def="v",
+                        spike_var="spike", op="qif_sfa_op", node_vars={'all/qif_sfa_op/eta': etas}, dt=dt,
                         spike_threshold=v_theta, spike_reset=-v_theta, float_precision="float64", clear=True)
 
 # wash out initial condition